PIP-1183 Upgrade to Spring Boot 4, Elasticsearch 9 and Jackson 3

+++ src/main/java/org/snomed/snowstorm/core/data/services/pojo/SimpleAggregation.java
@@ -12,7 +12,7 @@ public class SimpleAggregation extends Aggregation {
 	private final long count;
 
 	public SimpleAggregation(String name, String bucket, long count) {
-		super(name, Aggregate.of(a -> a.bucketMetricValue(bm -> bm.keys(bucket).value(count))));
+		super(name, Aggregate.of(a -> a.bucketMetricValue(bm -> bm.keys(bucket).value((double) count))));
 		this.bucket = bucket;
 		this.count = count;
 	}
